added clear_folder arg and code improvements

+++ prep_data.py
@@ -1,7 +1,7 @@
 """
-Author(s): Greg Holste (giholste@gmail.com),  UT Austin
-           Jonathan Burkow (burkowjo@msu.edu), Michigan State University
-Last Modified: 02/21/2022
+Author(s): Jonathan Burkow (burkowjo@msu.edu), Michigan State University
+           Greg Holste (giholste@gmail.com),  UT Austin
+Last Modified: 06/05/2023
 Description: Converts rib fracture data to proper format + directory structure for YOLOv5 training.
 
 Usage: python prep_data.py --data_dir path/to/root/data/directory \
@@ -9,20 +9,36 @@ Usage: python prep_data.py --data_dir path/to/root/data/directory \
                            --train_anno <train_csv_name>.csv \
                            --val_anno <val_csv_name>.csv \
                            --test_anno <test_csv_name>.csv \
-                           --symbolic
+                           --symbolic \
+                           --clear_folder
 """
 
 import argparse
 import os
 import shutil
 import time
+from pathlib import Path
 
 import cv2
 import pandas as pd
 from tqdm import tqdm
 
 
-def create_directories(out_dir: str) -> None:
+def parse_cmd_args():
+    parser = argparse.ArgumentParser()
+
+    parser.add_argument('--data_dir', type=str)
+    parser.add_argument('--save_dir', type=str)
+    parser.add_argument('--train_anno', type=str, default='train_annotations.csv')
+    parser.add_argument('--val_anno', type=str, default='val_annotations.csv')
+    parser.add_argument('--test_anno', type=str, default='test_annotations.csv')
+    parser.add_argument('--symbolic', action='store_true', default=False)
+    parser.add_argument('--clear_folder', action='store_true', default=False)
+
+    return parser.parse_args()
+
+
+def create_directories(out_dir: str, clear: bool = False) -> None:
     """
     Create directories for images and labels to be used for YOLOv5 training.
 
@@ -30,7 +46,7 @@ def create_directories(out_dir: str) -> None:
     ----------
     out_dir : path to root directory for all YOLOv5 data folders
     """
-    if os.path.isdir(out_dir):
+    if os.path.isdir(out_dir) and clear:
         shutil.rmtree(out_dir)
         print('Removed previous data directories.')
     os.mkdir(out_dir)
@@ -45,7 +61,7 @@ def create_directories(out_dir: str) -> None:
     print('Created all data directories.')
 
 
-def copy_images(annot_path: str, set_name: str, symbolic: bool = False) -> None:
+def copy_images(annot_path: str, set_name: str, save_dir: str, symbolic: bool = False) -> None:
     """
     Copy images from the main data directory to the YOLO directory structure; can also be symbolic
     links if file count or disk space needs to be considered.
@@ -65,12 +81,12 @@ def copy_images(annot_path: str, set_name: str, symbolic: bool = False) -> None:
         patient_id = img_path.split('/')[-1]
 
         if symbolic:
-            os.symlink(img_path, os.path.join(args.save_dir, 'images', set_name, patient_id))
+            os.symlink(img_path, os.path.join(save_dir, 'images', set_name, patient_id))
         else:
-            shutil.copy(img_path, os.path.join(args.save_dir, 'images', set_name))
+            shutil.copy(img_path, os.path.join(save_dir, 'images', set_name))
 
 
-def convert_to_yolo(annot_path: str, set_name: str) -> None:
+def convert_to_yolo(annot_path: str, set_name: str, save_dir: str) -> None:
     """
     Convert labels from RetinaNet CSV format into YOLO txt format.
 
@@ -84,6 +100,7 @@ def convert_to_yolo(annot_path: str, set_name: str) -> None:
     classes = ['fracture']
 
     for _, img_path in tqdm(enumerate(annot['img_path'].unique().tolist()), desc=f'Converting {set_name.title()} Labels', total=len(annot['img_path'].unique().tolist())):
+        patient_id = img_path.split('/')[-1].split('.')[0]
         if not os.path.isfile(img_path):
             print(f"IMAGE FOR {patient_id} NOT FOUND/NOT READ IN")
             continue
@@ -94,9 +111,7 @@ def convert_to_yolo(annot_path: str, set_name: str) -> None:
         if sub_df.isnull().values.any():
             continue
 
-        patient_id = img_path.split('/')[-1].split('.')[0]
-
-        with open(os.path.join(args.save_dir, 'labels', set_name, f"{patient_id}.txt"), 'w') as out:
+        with open(os.path.join(save_dir, 'labels', set_name, f"{patient_id}.txt"), 'w') as out:
             if img_path.endswith('\''):
                 img_path = img_path[:-1]  # remove last character
 
@@ -113,38 +128,30 @@ def convert_to_yolo(annot_path: str, set_name: str) -> None:
                 w = row['x2'] - row['x1']
                 h = row['y2'] - row['y1']
 
-                data = [cls_idx, x_c/img_w, y_c/img_h, w/img_w, h/img_h]
+                data = [cls_idx, x_c / img_w, y_c / img_h, w / img_w, h / img_h]
 
                 out.write(' '.join(str(s) for s in data) + '\n')
 
 
-def main(parse_args):
+def main():
     """Main Function"""
-    create_directories(parse_args.save_dir)
+    parse_args = parse_cmd_args()
 
-    copy_images(os.path.join(parse_args.data_dir, parse_args.train_anno), 'train', parse_args.symbolic)
-    copy_images(os.path.join(parse_args.data_dir, parse_args.val_anno), 'val', parse_args.symbolic)
-    copy_images(os.path.join(parse_args.data_dir, parse_args.test_anno), 'test', parse_args.symbolic)
+    create_directories(parse_args.save_dir, clear=parse_args.clear_folder)
 
-    convert_to_yolo(os.path.join(parse_args.data_dir, parse_args.train_anno), 'train')
-    convert_to_yolo(os.path.join(parse_args.data_dir, parse_args.val_anno), 'val')
-    convert_to_yolo(os.path.join(parse_args.data_dir, parse_args.test_anno), 'test')
+    copy_images(os.path.join(parse_args.data_dir, parse_args.train_anno), 'train', parse_args.save_dir, parse_args.symbolic)
+    copy_images(os.path.join(parse_args.data_dir, parse_args.val_anno), 'val', parse_args.save_dir, parse_args.symbolic)
+    copy_images(os.path.join(parse_args.data_dir, parse_args.test_anno), 'test', parse_args.save_dir, parse_args.symbolic)
+
+    convert_to_yolo(os.path.join(parse_args.data_dir, parse_args.train_anno), 'train', parse_args.save_dir)
+    convert_to_yolo(os.path.join(parse_args.data_dir, parse_args.val_anno), 'val', parse_args.save_dir)
+    convert_to_yolo(os.path.join(parse_args.data_dir, parse_args.test_anno), 'test', parse_args.save_dir)
 
 
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser()
-    parser.add_argument('--data_dir', type=str)
-    parser.add_argument('--save_dir', type=str)
-    parser.add_argument('--train_anno', type=str, default='train_annotations.csv')
-    parser.add_argument('--val_anno', type=str, default='val_annotations.csv')
-    parser.add_argument('--test_anno', type=str, default='test_annotations.csv')
-    parser.add_argument('--symbolic', action='store_true', default=False)
-
-    args = parser.parse_args()
-
-    print('\nStarting execution...')
+if __name__ == "__main__":
+    print(f"\n{'Starting execution: ' + Path(__file__).name:-^80}\n")
     start_time = time.perf_counter()
-    main(args)
-    end_time = time.perf_counter()
-    print('Done!')
-    print(f'Execution finished in {end_time - start_time:.3f} seconds.\n')
+    main()
+    elapsed = time.perf_counter() - start_time
+    print(f"\n{'Done!':-^80}")
+    print(f'Execution finished in {elapsed:.3f} seconds ({time.strftime("%-H hr, %-M min, %-S sec", time.gmtime(elapsed))}).\n')
